+++ srcs/casting/ray.c
@@ -6,16 +6,16 @@
 /*   By: gcrisp <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 13:25:45 by gcrisp            #+#    #+#             */
-/*   Updated: 2025/02/21 13:25:46 by gcrisp           ###   ########.fr       */
+/*   Updated: 2025/02/21 15:34:37 by gcrisp           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <casting.h>
 
-t_ray	*new_ray(double x, double y, double degs)
+t_ray	*new_ray(float x, float y, float degs)
 {
 	t_ray	*out;
-	double	rads;
+	float	rads;
 
 	out = malloc(sizeof(t_ray));
 	out->pos = new_point(x, y);
